+++ uploadcare/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.1
+
+- feat: progressive operation
+- feat: image format -  webp, auto
+
+
 ## 0.1.0
 
 - CdnPathBuilder